Use delta to avoid rewriting the LP at each BnB node.
Previously we wastefully recomputed conflicting columns and rows and
rewrote the full LP at each BnB node. Now, we keep the same LP and
update it with delta as necessary at each node.

Note: this still leaves some more room for optimization because at every
node we apply the full delta and then undo the entirety of delta. It
might be faster to keep a per-node delta